backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
pydub==0.25.1
requests==2.32.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0